Client simulator: use pool of macs with single crt/key
Make client simulator use single cert/key pair with mac
being taken from predefind macs list. This makes it easier
to use simulator, as it doesn't depend upon a big list
of certificates that are expected to match simulated MAC
and so on.

Fixup some python imports to not use relative import,
but rather istruct makefile to pass python path var
that includes relative path.
This is a change required for following commits that
bring automated testing using this simulator.

Signed-off-by: Oleksandr Mazur <oleksandr.mazur@plvision.eu>