CGW: provide (optional) kafka routing key in some of replies
The provided kafka key in (some of the) replies can be used
by NB services to talk <directly> to _this_ local shard
that replied, effectively using kafka partitioning to
not rely on the relaying mechanism.
The keys list is calculated upon each kafka partition rebalancing,
if the partition number changes.

Signed-off-by: Oleksandr Mazur <oleksandr.mazur@plvision.eu>