CGW: rework partial infra add/del returned infras list
Change CGW behavior to return only 'failed_infras' in both
infra add/del partial or completely failed responses.

This tweaks the previous redundant data that was returned in the
response - list of <all> macs in case of a successfull completion,
which is unnecessary - it's already clear that request succeeded,
and actions were made over supplied infra list.

Signed-off-by: Oleksandr Mazur <oleksandr.mazur@plvision.eu>