CGW: ConnectionServer: prevent redis access race condition on init
Whenever connserver is being created, the NB API loop
task is created earlier than redis (remote discovery)
is properly initialized and remote cgw map / gid cache
is synced.
This causes the very few initial redis calls to fail,
which set the redis health to not-healty.
As a result, this transition might kill the underlying
pod as it's redeemed unhealthy - dead.

Signed-off-by: Oleksandr Mazur <oleksandr.mazur@plvision.eu>